Feat: 등록 후 새로고침 되는 훅 추가
