+++ src/pages/PlaylistForm/hooks/useSubmitPlaylist.ts
@@ -1,13 +1,7 @@
 import { useState } from 'react'
 
 import { PlaylistFormValues } from '@/pages//PlaylistForm/model/types'
-import {
-  useCreatePlaylist,
-  useCreatePlaylistItems,
-  useUpdateThumbnailUrl,
-  useUploadThumbnail,
-} from '@/pages/PlaylistForm/queries/usePlaylistQuery'
-import { useUserStore } from '@/shared/store/userStore'
+import { useSubmitPlaylistMutation } from '@/pages/PlaylistForm/queries/usePlaylistQuery'
 
 interface UseSubmitPlaylistProps {
   onSuccess?: () => void
@@ -25,15 +19,8 @@ interface FormattedPlaylistData {
 }
 
 export const useSubmitPlaylist = ({ onSuccess, onError }: UseSubmitPlaylistProps = {}) => {
-  const [isSubmitting, setIsSubmitting] = useState(false)
-  const [submitError, setSubmitError] = useState<Error | null>(null)
   const [formattedData, setFormattedData] = useState<FormattedPlaylistData | null>(null)
-
-  // Playlist 생성 및 업로드 관련 mutations
-  const createPlaylistMutation = useCreatePlaylist()
-  const createPlaylistItemsMutation = useCreatePlaylistItems()
-  const uploadThumbnailMutation = useUploadThumbnail()
-  const updateThumbnailUrlMutation = useUpdateThumbnailUrl()
+  const submitPlaylistMutation = useSubmitPlaylistMutation()
 
   const formatPlaylistData = (data: PlaylistFormValues): FormattedPlaylistData => {
     return {
@@ -49,58 +36,7 @@ export const useSubmitPlaylist = ({ onSuccess, onError }: UseSubmitPlaylistProps
 
   const submitPlaylist = async (data: PlaylistFormValues) => {
     try {
-      setIsSubmitting(true)
-      setSubmitError(null)
-
-      const profileId = useUserStore.getState().profileId
-      if (!profileId) throw new Error('사용자 정보를 찾을 수 없습니다.')
-
-      // 1. 플레이리스트 생성
-      const newPlaylist = await createPlaylistMutation.mutateAsync({
-        title: data.title,
-        description: data.description || null,
-        profile_id: profileId,
-        is_public: data.isPublic,
-        hashtag: data.hashtags.length ? data.hashtags : undefined,
-        thumbnail_url: null,
-      })
-
-      // 2. 플레이리스트 아이템 생성
-      await createPlaylistItemsMutation.mutateAsync({
-        playlistId: newPlaylist.id,
-        videos: data.videos,
-      })
-
-      // 3. 썸네일 처리
-      let thumbnailUrl = ''
-      if (data.thumbnail) {
-        try {
-          // 썸네일 업로드
-          thumbnailUrl = await uploadThumbnailMutation.mutateAsync({
-            file: data.thumbnail,
-            userId: profileId,
-            playlistId: newPlaylist.id,
-          })
-          // 썸네일 URL 업데이트
-          await updateThumbnailUrlMutation.mutateAsync({
-            playlistId: newPlaylist.id,
-            thumbnailUrl,
-          })
-        } catch (thumbError) {
-          setSubmitError(new Error('썸네일 업로드 또는 URL 업데이트 중 오류가 발생했습니다.'))
-          throw thumbError
-        }
-      } else if (data.videos.length > 0) {
-        try {
-          await updateThumbnailUrlMutation.mutateAsync({
-            playlistId: newPlaylist.id,
-            thumbnailUrl: data.videos[0].thumbnailUrl,
-          })
-        } catch (thumbError) {
-          setSubmitError(new Error('기본 썸네일 URL 업데이트 중 오류가 발생했습니다.'))
-          throw thumbError
-        }
-      }
+      await submitPlaylistMutation.mutateAsync(data)
 
       // 데이터 형식 변환 (표시용)
       const formatted = formatPlaylistData(data)
@@ -110,11 +46,8 @@ export const useSubmitPlaylist = ({ onSuccess, onError }: UseSubmitPlaylistProps
       return true
     } catch (error) {
       const err = error instanceof Error ? error : new Error('알 수 없는 오류가 발생했습니다.')
-      setSubmitError(err)
       onError?.(err)
       return false
-    } finally {
-      setIsSubmitting(false)
     }
   }
 
@@ -150,8 +83,8 @@ export const useSubmitPlaylist = ({ onSuccess, onError }: UseSubmitPlaylistProps
   return {
     submitPlaylist,
     validatePlaylist,
-    isSubmitting,
-    submitError,
+    isSubmitting: submitPlaylistMutation.isPending,
+    submitError: submitPlaylistMutation.error as Error | null,
     formattedData,
   }
 }
